caop: bitmask fast-rejection, popcnt sign factor, int bit_length
Adds three optimizations to the CAOP method 0 mult kernel, collectively
delivering ~22% wall-time reduction on a 1.6M-basis 30-orbital benchmark
at 4 ranks (68.7s → 53.7s davidson time).

Bitmask fast-rejection (generalop.h, mult.h):
- GeneralOp::PrecomputeMasks() builds per-term m1_ (creation targets,
  must be occupied in bra) and m2_ (annihilation targets, must be
  unoccupied), stored as mutable det_vector<size_t>.
- m2_ is masked against m1_ at construction to handle orbitals that
  appear in both creation and annihilation sets of the same term.
- mult() calls PrecomputeMasks lazily on first invocation, then checks
  (~vb & m1_[n]) || (vb & m2_[n]) before entering the k-loop.
  Rejects ~94% of 2-body terms before any bit manipulation.
- GeneralOp friend declaration updated to match new int bit_length.

Popcnt sign factor (bit_manipulation.h):
- bit_string_sign_factor replaces nested O(bit_length * det_words)
  loops with __builtin_popcountll: parity over full words before r,
  then partial word masked to bits below x.
- Arguments x and r changed from size_t to int; loop variable k int.

int bit_length throughout (mult.h, davidson.h, lanczos.h, sbdiag.h):
- bit_length parameter changed from size_t to int in mult(), Davidson(),
  Lanczos(), and the local variable in caop::diag().
- Eliminates size_t/int mixed arithmetic in the inner loop (q/r/x now
  plain int division/modulo); static_cast<size_t> on fops_[k].q_ removed.

+++ include/sbd/caop/basic/davidson.h
@@ -39,7 +39,7 @@ namespace sbd {
   void Davidson(const std::vector<ElemT> & hii,
 		std::vector<ElemT> & W,
 		const det_vector<size_t> & bs,
-		const size_t bit_length,
+		const int bit_length,
 		const std::vector<int> & slide,
 		const GeneralOp<ElemT> & Ham,
 		bool sign,
